Centralize validation logic in common folder.
