+++ packages/babel-core/src/config/index.js
@@ -3,7 +3,7 @@
 import type Plugin from "./plugin";
 import manageOptions from "./option-manager";
 
-export type { InputOptions } from "./options";
+export type { InputOptions } from "./validation/options";
 
 export type ResolvedConfig = {
   options: Object,
