+++ packages/babel-core/src/config/validation/option-assertions.js
