+++ packages/babel-core/src/config/validation/options.js
