+++ packages/babel-core/src/config/validation/removed.js
